refactor(routing): rename global to routing
